requirements.txt
fastapi
uvicorn
pydantic
openai>=1.0.0
nest_asyncio
requests
python-dotenv
